added various conversions

+++ crates/brontes-database/src/lib.rs
@@ -22,7 +22,7 @@ pub struct Metadata {
     pub relay_timestamp:        u64,
     pub p2p_timestamp:          u64,
     pub proposer_fee_recipient: Address,
-    pub proposer_mev_reward:    u64,
+    pub proposer_mev_reward:    u128,
     pub cex_quotes:             PriceGraph,
     /// Best ask at p2p timestamp
     pub eth_prices:             Rational,
@@ -145,7 +145,7 @@ impl Metadata {
         relay_timestamp: u64,
         p2p_timestamp: u64,
         proposer_fee_recipient: Address,
-        proposer_mev_reward: u64,
+        proposer_mev_reward: u128,
         cex_quotes: PriceGraph,
         eth_prices: Rational,
         mempool_flow: HashSet<TxHash>,
@@ -165,8 +165,8 @@ impl Metadata {
 }
 
 impl Metadata {
-    pub fn get_gas_price_usd(&self, gas_used: u64) -> Rational {
-        let gas_used_rational = Rational::from_unsigneds(gas_used, 10u64.pow(18));
+    pub fn get_gas_price_usd(&self, gas_used: u128) -> Rational {
+        let gas_used_rational = Rational::from_unsigneds(gas_used, 10u128.pow(18));
 
         &self.eth_prices * gas_used_rational
     }
